requirements.txt
# requirements.txt

pyspark==3.5.0
kafka-python==2.0.2
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.0
pandas==2.2.1
